Commit code generated from protobuf schemas
This change enables the crates to be compiled from source without
requiring the installation of `protoc` (if, for example, the crates
were published to a public crate repository for consumption by other
users).

The change introduces a new intermediate crate `aurae-proto` that
is depended upon by both `auraed` and `auraescript` rather than
having `tonic-build` re-compile the protobuf into each of the crates
independently. Generated code is checked in so that it is available
without needing any `build.rs` to invoke `npm` or other external
programs at crate compile-time.

In addition, this change fixes the issue with TypeScript and Rust
protoc-generated code not agreeing on the naming convention being
used for fields when serialized through a JSON interface. By using
`pbjson`, we correctly add the appropriate serializers/deserializers
to ensure that we get the correct cannonical JSON serialization for
protobuf types.

Generating these files now relies on a program `buf` which invokes
some remote plugins to generate Rust and TypeScript code. The `Makefile`
has been updated to provide two new commands: `proto` and `proto-lint`.
`proto` runs `buf generate` to generate code from the protobuf
schemas, while `lint` runs `buf lint` to provide linting recommendations
for the protobuf definitions. The linting is configurable, but I did not
update the schemas to match the `buf` defaults, nor try to configure
the linting to match the current _status quo_.

`buf` also provides a `buf format` command, which can be used to
ensure that the protobuf definition files follow a standard format,
similar to `rustfmt`, but this also has not been incorporated or run.

Fixes #128