Remove docker builds in ci & dependency linting
It's not clear what specific value the docker builds provide currently. It's an
interesting use case to revists in the future. As for cargo-deny, it's too
early in the lifecycle of the project to be this hardcore about our dependencies.
cargo-deny is definitely something we want to readd to ci as we get closer to a
1.0 release candidate